Moved configure_s3_access to utils.aws.
This method is more close to aws than to rio, also current documentation point
to uitils.aws. location.

Still keeping forwarding function for backwards compatibility though, currently
both locations are used in dea-notebooks repo.
