+++ datacube/utils/aws/__init__.py
@@ -6,18 +6,18 @@
 Helper methods for working with AWS
 """
 import os
+import time
+import functools
 import botocore
 import botocore.session
 from botocore.credentials import Credentials, ReadOnlyCredentials
 from botocore.session import Session
-import time
 from urllib.request import urlopen
 from urllib.parse import urlparse
 from sqlalchemy.engine.url import URL
 
 from typing import Optional, Dict, Tuple, Any, Union, IO
 from datacube.utils.generic import thread_local_cache
-from ..rio import configure_s3_access
 
 ByteRange = Union[slice, Tuple[int, int]]       # pylint: disable=invalid-name
 MaybeS3 = Optional[botocore.client.BaseClient]  # pylint: disable=invalid-name
@@ -450,3 +450,56 @@ def obtain_new_iam_auth_token(url: URL, region_name: str = "auto", profile_name:
     client = session.client("rds", region_name=region_name)
     return client.generate_db_auth_token(DBHostname=url.host, Port=url.port, DBUsername=url.username,
                                          Region=region_name)
+
+
+def configure_s3_access(profile=None,
+                        region_name="auto",
+                        aws_unsigned=False,
+                        requester_pays=False,
+                        cloud_defaults=True,
+                        client=None,
+                        **gdal_opts):
+    """ Credentialize for S3 bucket access or configure public access.
+
+    This function obtains credentials for S3 access and passes them on to
+    processing threads, either local or on dask cluster.
+
+
+    .. note::
+
+       if credentials are STS based they will eventually expire, currently
+       this case is not handled very well, reads will just start failing
+       eventually and will never recover.
+
+    :param profile:        AWS profile name to use
+    :param region_name:    Default region_name to use if not configured for a given/default AWS profile
+    :param aws_unsigned:   If ``True`` don't bother with credentials when reading from S3
+    :param requester_pays: Needed when accessing requester pays buckets
+
+    :param cloud_defaults: Assume files are in the cloud native format, i.e. no side-car files, disables
+                           looking for side-car files, makes things faster but won't work for files
+                           that do have side-car files with extra metadata.
+
+    :param client:         Dask distributed ``dask.Client`` instance, if supplied apply settings on the
+                           dask cluster rather than locally.
+    :param gdal_opts:      Any other option to pass to GDAL environment setup
+
+    :returns: credentials object or ``None`` if ``aws_unsigned=True``
+    """
+    from ..rio import set_default_rio_config
+
+    aws, creds = get_aws_settings(profile=profile,
+                                  region_name=region_name,
+                                  aws_unsigned=aws_unsigned,
+                                  requester_pays=requester_pays)
+
+    if client is None:
+        set_default_rio_config(aws=aws, cloud_defaults=cloud_defaults, **gdal_opts)
+    else:
+        client.register_worker_callbacks(
+            functools.partial(set_default_rio_config,
+                              aws=aws,
+                              cloud_defaults=cloud_defaults,
+                              **gdal_opts))
+
+    return creds
